Refactor URL handler
